Create Hafiz Uvesh
Create a 3D illustration of an animated character sitting casually on top of a social media logo "Instagram
". The character must wear casual modern clothing such as jeans jacket and sneakers shoes. The background of the image is a social media profile page with a user name "Official_hafiz_uvesh9" and a profile picture that match.